quicksecscan/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery[redis]==5.3.4
redis==5.0.1
stripe==7.4.0
dnspython==2.4.2
jinja2==3.1.2
weasyprint==60.1
boto3==1.29.7
sendgrid==6.10.0
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0
PyJWT==2.8.0

